server_fastapi/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pillow==11.0.0
httpx==0.28.1
replicate==1.0.0
supabase==2.10.0
python-dotenv==1.0.1
pydantic-settings==2.6.1
